backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# MongoDB
motor==3.3.2
pymongo==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Image Processing
Pillow==10.1.0
python-magic==0.4.27

# HTTP Client
httpx==0.25.2
aiofiles==23.2.0

# Environment Management
python-dotenv==1.0.0

# Logging and Monitoring
loguru==0.7.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# No need for python-cors
# python-cors==1.0.1